apply strongswan patch changing psk lookup order
Change lookup order for PSKs to be used with an incoming IKEv1
connection.

Before, the remote peer's IP address was matched against
the entries in IP secrets. This never matched FQDN (domain names) but
always matched an "%any" entry (e.g., if roadwarriors were
configured for PSK alongside a DynDNS-derived site-to-site connection).

Now, strongswan first matches the remote peer's IP address with the
configuration from ipsec.conf (applying DNS name resolution when
updating he config!) and if a suitable configuration is found uses the
identities stored there to match the PSK entry (which can also be a
hostname). Only if unsuccessful (e.g., if the IP can only be matched to
"%any", yielding no matched PSK), the lookup by IP is used, matching
the any entry for the roadwarrior.

fixes #7845